requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
chardet>=5.0.0
rapidfuzz>=3.0.0
python-dateutil>=2.8.0
pymannkendall>=1.4.3
rich>=13.0.0
statsmodels>=0.14.0
pyyaml>=6.0.0
openpyxl>=3.1.0
pytest>=7.0.0
